+++ packages/@inrealtime/react/src/createRealtimeContext.tsx
@@ -14,11 +14,11 @@ import { PresenceClient } from './core'
 import {
   RealtimeConnectionStatus,
   RealtimeDocumentStatus,
+  RealtimeOptions,
   useRealtime,
-  UseRealtimeOptions,
 } from './useRealtime'
 
-type RealtimeContextProps<TRealtimeState, TRealtimePresenceData> = {
+export type RealtimeContextProps<TRealtimeState, TRealtimePresenceData> = {
   documentStatus: RealtimeDocumentStatus
   connectionStatus: RealtimeConnectionStatus
   useStore: UseStore<TRealtimeState>
@@ -35,7 +35,7 @@ type RealtimeContextProps<TRealtimeState, TRealtimePresenceData> = {
 
 type RealtimeProviderProps = {
   children: React.ReactNode
-} & UseRealtimeOptions
+} & RealtimeOptions
 
 type RealtimeContextCollection<TRealtimeState, TRealtimePresenceData> = {
   RealtimeProvider(props: RealtimeProviderProps): JSX.Element