+++ packages/@inrealtime/react/src/index.ts
@@ -1,4 +1,13 @@
 export type { DocumentPatch } from './channels/document/store/types'
+export type { AutosaveOption } from './config'
+export type { GetRealtimeAuthToken } from './core'
 export type { PresenceClient } from './core/types/presence'
-export { createRealtimeContext } from './createRealtimeContext'
-export { RealtimeConnectionStatus, RealtimeDocumentStatus, useRealtime } from './useRealtime'
+export { createRealtimeContext, RealtimeContextProps } from './createRealtimeContext'
+export { useAutosave } from './useAutosave'
+export {
+  RealtimeConnectionStatus,
+  RealtimeDocumentStatus,
+  RealtimeOptions,
+  RealtimePackageOptions,
+  useRealtime,
+} from './useRealtime'