+++ packages/tests/@realtimetests/react/CHANGELOG.md
@@ -1,5 +1,83 @@
 # @realtimetests/react
 
+## 0.0.18
+
+### Patch Changes
+
+- Test
+
+## 0.0.17
+
+### Patch Changes
+
+- Test
+
+## 0.0.16
+
+### Patch Changes
+
+- Testing autosave new hooks
+
+## 0.0.15
+
+### Patch Changes
+
+- Fixes to autosave
+
+## 0.0.14
+
+### Patch Changes
+
+- Updated autosave options"
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated autosave options
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated jwt token
+
+## 0.0.11
+
+### Patch Changes
+
+- Added more types expose
+
+## 0.0.10
+
+### Patch Changes
+
+- More exposing of types
+
+## 0.0.9
+
+### Patch Changes
+
+- Exposed realtime options"
+
+## 0.0.8
+
+### Patch Changes
+
+- Exposed get realtime auth token
+
+## 0.0.7
+
+### Patch Changes
+
+- Added error messages in auth in realtime
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated how websockets live, specifically for iOS safari, updated ping pong strat
+
 ## 0.0.5
 
 ### Patch Changes